Improvements

+++ Server/system.DataAccess/Entites/UserEntity.cs
@@ -2,17 +2,12 @@
 
 public class UserEntity
 {
-    public int Id { get; set; }
-
+    public long Id { get; set; }
     public int RoleId { get; set; }
-
     public string Login { get; set; } = string.Empty;
-
     public string PasswordHash { get; set; } = string.Empty;
 
     public ClientEntity? Client { get; set; }
-
     public WorkerEntity? Worker { get; set; }
-
     public RoleEntity? Role { get; set; }
 }
